+++ lib/Command.ts
