+++ lib/cluster/index.ts
@@ -2,8 +2,8 @@ import { EventEmitter } from "events";
 import * as commands from "redis-commands";
 import { AbortError, RedisError } from "redis-errors";
 import asCallback from "standard-as-callback";
-import Pipeline from "../pipeline";
-import Command from "../command";
+import Pipeline from "../Pipeline";
+import Command from "../Command";
 import ClusterAllFailedError from "../errors/ClusterAllFailedError";
 import Redis from "../Redis";
 import ScanStream from "../ScanStream";
